Move Diagnostic type into its own module
