+++ polar-core/src/diagnostic.rs
@@ -1,0 +1,6 @@
+use super::error::PolarError;
+
+pub enum Diagnostic {
+    Error(PolarError),
+    Warning(String),
+}
